refactor: 런타임에 플랫폼에 대응되는 인스턴스를 통해 로직이 실행되도록 변경
