+++ ShareCode/ShareCode/src/Platform/PlatformHandler.cs
@@ -1,8 +1,10 @@
-﻿namespace ShareCode.src.Platform
+﻿using System.Collections.ObjectModel;
+
+namespace ShareCode.src.Platform
 {
     public abstract class PlatformHandler
     {
-        public abstract Task OnOpenButtonClicked(object? sender, EventArgs e);
-        public abstract Task OnExportButtonClicked(object? sender, EventArgs e);
+        public abstract Task OnOpenButtonClicked(ContentPage page, ObservableCollection<FileItem> originFileList, ObservableCollection<FileItem> fileList);
+        public abstract Task OnExportButtonClicked(ContentPage page, ObservableCollection<FileItem> fileList);
     }
 }
